Factor out requirement components in RequirementFulfillment (3/n)
